Fixed null return in entity provider.
Changed average score calculation to be based on gradings, not
submissions. Previously I was calculated the average based on the
median between the highest and lowest scores - taking into account
the possibility that a student could submit many times per grading
and that the number of submissions was not a reliable divisor. I've
now added an extra column to store the number of actual gradings and
that's what is now used in the average score calculation.